MdeModulePkg: CodeQL Fixes.
Includes changes across the repo for the following CodeQL rules:
- cpp/comparison-with-wider-type
- cpp/overflow-buffer
- cpp/redundant-null-check-param
- cpp/uselesstest

+++ MdeModulePkg/Universal/Disk/PartitionDxe/Mbr.c
@@ -359,7 +359,7 @@ PartitionInstallMbrChildHandles (
       if (ExtMbrStartingLba == 0) {
         break;
       }
-    } while (ExtMbrStartingLba  < ParentHdDev.PartitionSize);
+    } while ((UINT64)ExtMbrStartingLba  < ParentHdDev.PartitionSize);
   }
 
 Done:
